backend/pyproject.toml
[tool.poetry]
name = "black-edge"
version = "0.1.0"
description = "Quantitative arbitrage engine for prediction markets"
authors = ["Black Edge Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
web3 = ">=7.0.0"
numpy = "^1.26.0"
scipy = "^1.12.0"
cvxpy = "^1.4.0"
redis = "^5.0.0"
httpx = ">=0.27.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
websockets = "^12.0"
firebase-admin = "^6.4.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
